feat(composable): ajoute le composable useTabs
Pour gérer facilement l’onglet à afficher et le sens de l’animation
