+++ src/components/DsfrAccordion/DsfrAccordion.vue
@@ -2,7 +2,7 @@
 import { computed, onMounted, watch } from 'vue'
 
 import { getRandomId } from '../../utils/random-utils'
-import { useCollapsable } from '../../composables.js'
+import { useCollapsable } from '../../composables'
 import type { DsfrAccordionProps } from './DsfrAccordion.types'
 
 export type { DsfrAccordionProps }
