+++ src/composables/useTabs.stories.mdx
@@ -1,0 +1,58 @@
+import { Meta } from '@storybook/blocks';
+
+<Meta title="Composables/useTabs" />
+
+# `useTabs`
+
+Ce composable permet de gérer simplement les props de [DsfrTabs](/?path=/docs/composants-dsfrtabs--docs) et de ses enfants pour gérer l’affichage du contenu de l’onglet (grâce à la prop `selected` de `DsfrTabContent`) et le sens de l’animation (grâce à la prop `asc` sur `DsfrTabContent`).
+
+## Utilisation avancée (Script setup)
+
+```html
+<template>
+   <DsfrTabs
+     ref="tabs"
+     :tab-list-name="tabListName"
+     :tab-titles="tabTitles"
+     :initial-selected-index="selectedTabIndex"
+     @select-tab="selectTab"
+   >
+     <DsfrTabContent
+       panel-id="tab-content-0"
+       tab-id="tab-0"
+       :selected="selectedTabIndex === 0"
+       :asc="asc"
+     >
+       <div>Contenu 1 avec d'autres composants</div>
+     </DsfrTabContent>
+     <DsfrTabContent
+       panel-id="tab-content-1"
+       tab-id="tab-1"
+       :selected="selectedTabIndex === 1"
+       :asc="asc"
+     >
+       <div>Contenu 2 avec d'autres composants</div>
+     </DsfrTabContent>
+   </DsfrTabs>
+</template>
+
+<script setup lang="ts">
+import { useTabs, DsfrTabs } from '@gouvminint/vue-dsfr'
+
+const { asc, selectedTabIndex, selectTab } = useTabs(true,0)
+
+const tabListName = "Onglets"
+
+const tabTitles: { title: string; tabId: string }[] = [
+  {
+    title: 'Informations générales',
+    tabId: 'information-generale',
+  },
+  {
+    title: 'Document(s)',
+    tabId: 'documents',
+  },
+]
+
+</script>
+```
