+++ src/index.ts
@@ -25,6 +25,6 @@ export default {
 
 export * from './components/index.js'
 export * from './utils/random-utils'
-export * from './composables.js'
+export * from './composables/index'
 export * from './common-types.js'
 export * as icons from './icons.js'
